// simulation_servers/deno/src/tools/third_party_modules/deno_mongo@0.31.1/src/types.ts
export interface Server {
  host: string;

  port: number;

  domainSocket?: string;
}

export interface ConnectOptions {
  compression?: string[];

  caCert?: string | URL;

  privateKey?: string | URL;

  certChain?: string | URL;

  keyFilePassword?: string;

  tls?: boolean;

  safe?: boolean;

  credential?: Credential;

  db: string;

  servers: Server[];

  retryWrites?: boolean;

  appname?: string;
}
// ...
export interface Credential {
  /**
   * The username to authenticate with. This applies to all mechanisms but may be omitted when authenticating via MONGODB-X509.
   */
  username?: string;

  /**
   * The password to authenticate with. This does not apply to all mechanisms.
   */
  password?: string;

  /**
   * The database used to authenticate. This applies to all mechanisms and defaults to "admin" in SCRAM authentication mechanisms and "$external" for GSSAPI, MONGODB-X509 and PLAIN.
   */
  db?: string;

  /**
   * Which authentication mechanism to use. If not provided, one will be negotiated with the server.
   */
  mechanism?: 'SCRAM-SHA-1' | 'SCRAM-SHA-256' | 'MONGODB-X509';
}